@@ -76,8 +76,8 @@
         <maven.version.rules>file://${project.basedir}/config/versions-rule-set.xml</maven.version.rules>
         <pmd.version>6.51.0</pmd.version>
         <sb-contrib.version>7.4.7</sb-contrib.version>
-        <spring.version>5.3.23</spring.version>
-        <spring-boot.version>2.7.5</spring-boot.version>
+        <spring.version>6.0.2</spring.version>
+        <spring-boot.version>3.0.0</spring-boot.version>
     </properties>
 
     <dependencyManagement>
@@ -110,7 +110,7 @@
             <dependency>
                 <groupId>org.slf4j</groupId>
                 <artifactId>slf4j-api</artifactId>
-                <version>2.0.3</version>
+                <version>2.0.5</version>
             </dependency>
         </dependencies>
     </dependencyManagement>
@@ -138,9 +138,9 @@
             <version>${spring-boot.version}</version>
         </dependency>
         <dependency>
-            <groupId>javax.inject</groupId>
-            <artifactId>javax.inject</artifactId>
-            <version>1</version>
+            <groupId>jakarta.inject</groupId>
+            <artifactId>jakarta.inject-api</artifactId>
+            <version>2.0.1</version>
         </dependency>
         <dependency>
             <groupId>org.junit.jupiter</groupId>
@@ -279,7 +279,7 @@
                 <plugin>
                     <groupId>org.owasp</groupId>
                     <artifactId>dependency-check-maven</artifactId>
-                    <version>7.3.0</version>
+                    <version>7.3.2</version>
                     <configuration>
                         <dataDirectory>${user.dir}/../dependency-check-data</dataDirectory>
                         <cveValidForHours>12</cveValidForHours>
@@ -395,7 +395,7 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-install-plugin</artifactId>
-                <version>3.0.1</version>
+                <version>3.1.0</version>
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
@@ -429,6 +429,7 @@
                                         <exclude>META-INF/DEPENDENCIES</exclude>
                                         <exclude>META-INF/spring.tooling</exclude>
                                         <exclude>META-INF/versions/9/module-info.class</exclude>
+                                        <exclude>META-INF/spring/aot.factories</exclude>
                                     </excludes>
                                 </filter>
                             </filters>
