@@ -20,7 +20,7 @@ import java.io.BufferedWriter;
 import java.io.IOException;
 import java.util.Locale;
 
-import javax.inject.Inject;
+import jakarta.inject.Inject;
 
 import org.springframework.context.MessageSource;
 import org.springframework.stereotype.Component;
